move to App.vue
