+++ src/App.vue
@@ -3,8 +3,9 @@
 </template>
 
 <script type="ts">
-import { defineComponent } from "vue";
+import { computed, defineComponent, watch } from "vue";
 import { useStore, SAVE_PROJECT_FILE, LOAD_PROJECT_FILE } from "@/store";
+import { UI_LOCKED, UPDATE_MENU } from "@/store/ui";
 import {
   GENERATE_AND_SAVE_ALL_AUDIO,
   IMPORT_FROM_FILE,
@@ -20,6 +21,17 @@ export default defineComponent({
   name: "App",
   setup() {
     const store = useStore();
+    const uiLocked = computed(() => store.getters[UI_LOCKED]);
+    const isEngineReady = computed(() => store.state.isEngineReady);
+
+    const updateMenu = () => {
+      store.dispatch(UPDATE_MENU, {
+        uiLocked: uiLocked.value || !isEngineReady.value,
+      });
+    };
+    watch(uiLocked, updateMenu);
+    watch(isEngineReady, updateMenu);
+    updateMenu();
 
     window.electron.onReceivedIPCMsg(IPC_GENERATE_AND_SAVE_ALL_AUDIO, () => store.dispatch(GENERATE_AND_SAVE_ALL_AUDIO, {}));
     window.electron.onReceivedIPCMsg(IPC_IMPORT_FROM_FILE, () => store.dispatch(IMPORT_FROM_FILE, {}));
